Only consider the subject type when determining whether to include non-numerical properties and symbols.
Fixes the test where an arguments object is used as the RHS of a `<array-like> to satisfy <array-like>`
on node.js 4 and 5 where arguments objects have some weird ArrayValues symbol.

This needs to be thought through so we know it does the right thing when the RHS is a subtype
of array-like, but I think this is the way to go.